refactor(Card): move dueString and nextDue
Not in libanki
