+++ AnkiDroid/src/test/java/com/ichi2/libanki/CardTest.kt
@@ -25,7 +25,6 @@ import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.robolectric.annotation.Config
 import java.util.*
 import kotlin.test.assertNotNull
 
@@ -214,101 +213,5 @@ class CardTest : JvmTest() {
             val ord = card.ord
             assumeThat(ords, hasItemInArray(ord))
         }
-    }
-
-    @SuppressLint("DirectCalendarInstanceUsage")
-    @Test
-    @Config(qualifiers = "en")
-    fun nextDueTest() {
-        // Test runs as the 7th of august 2020, 9h00
-        val n = addNoteUsingBasicModel("Front", "Back")
-        val c = n.firstCard()
-        val decks = col.decks
-        val cal = Calendar.getInstance()
-        cal[2021, 2, 19, 7, 42] = 42
-        val id = cal.timeInMillis / 1000
-
-        // Not filtered
-        c.type = Consts.CARD_TYPE_NEW
-        c.due = 27L
-        c.queue = Consts.QUEUE_TYPE_MANUALLY_BURIED
-        assertEquals("27", c.nextDue())
-        assertEquals("(27)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SIBLING_BURIED
-        assertEquals("27", c.nextDue())
-        assertEquals("(27)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SUSPENDED
-        assertEquals("27", c.nextDue())
-        assertEquals("(27)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_NEW
-        c.due = 27L
-        assertEquals("27", c.nextDue())
-        assertEquals("27", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_PREVIEW
-        assertEquals("27", c.nextDue())
-        assertEquals("27", c.dueString())
-        c.type = Consts.CARD_TYPE_LRN
-        c.due = id
-        c.queue = Consts.QUEUE_TYPE_MANUALLY_BURIED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SIBLING_BURIED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SUSPENDED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_LRN
-        assertEquals("3/19/21", c.nextDue())
-        assertEquals("3/19/21", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_PREVIEW
-        assertEquals("", c.nextDue())
-        assertEquals("", c.dueString())
-        c.type = Consts.CARD_TYPE_REV
-        c.due = 20
-        //  Since tests run the 7th of august, in 20 days we are the 27th of august 2020
-        c.queue = Consts.QUEUE_TYPE_MANUALLY_BURIED
-        assertEquals("8/27/20", c.nextDue())
-        assertEquals("(8/27/20)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SIBLING_BURIED
-        assertEquals("8/27/20", c.nextDue())
-        assertEquals("(8/27/20)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SUSPENDED
-        assertEquals("8/27/20", c.nextDue())
-        assertEquals("(8/27/20)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_REV
-        assertEquals("8/27/20", c.nextDue())
-        assertEquals("8/27/20", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_PREVIEW
-        assertEquals("", c.nextDue())
-        assertEquals("", c.dueString())
-        c.type = Consts.CARD_TYPE_RELEARNING
-        c.due = id
-        c.queue = Consts.QUEUE_TYPE_MANUALLY_BURIED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SIBLING_BURIED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SUSPENDED
-        assertEquals("", c.nextDue())
-        assertEquals("()", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_LRN
-        c.due = id
-        assertEquals("3/19/21", c.nextDue())
-        assertEquals("3/19/21", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_PREVIEW
-        assertEquals("", c.nextDue())
-        assertEquals("", c.dueString())
-
-        // Dynamic deck
-        val dyn = decks.newFiltered("dyn")
-        c.oDid = c.did
-        c.did = dyn
-        assertEquals("(filtered)", c.nextDue())
-        assertEquals("(filtered)", c.dueString())
-        c.queue = Consts.QUEUE_TYPE_SIBLING_BURIED
-        assertEquals("(filtered)", c.nextDue())
-        assertEquals("((filtered))", c.dueString())
     }
 }
